pyproject.toml
[project]
name = "detectflow-backend"
version = "0.9.2"
description = "DetectFlow Backend API for managing pipelines, rules, and repositories"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Vladyslav Gumeniuk", email = "vladyslav.gumeniuk@socprime.com"},
    {name = "Viktor Hrebeniuk", email = "vh@socprime.com"},
    {name = "Vitalii Uslystyi", email = "vu@socprime.com"},

]

dependencies = [
    "schema-parser",
    "aiokafka>=0.12.0",
    "alembic>=1.14.0",
    "apscheduler>=3.11.1",
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.0",
    "confluent-kafka>=2.8.0",
    "fastapi>=0.121.3",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "kubernetes>=34.1.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn>=0.38.0",
]

[tool.uv.sources]
schema-parser = { git = "https://github.com/socprime/detectflow-parser.git", tag = "v0.2.3" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps"]

[tool.uv]
# UV best practices
managed = true              # Mark as uv-managed project
package = true              # Enable package management
default-groups = ["dev"]    # Auto-install dev dependencies

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "pre-commit>=4.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]

[tool.ruff]
line-length = 120
target-version = "py312"

# Ruff linting rules
lint.select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]

lint.ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (FastAPI Depends pattern)
    "UP046",  # Generic subclass (valid pattern for base DAO)
    "B904",   # raise-without-from-inside-except
]

# Format settings
format.quote-style = "double"
format.indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning",
]
